还原结构体定义

+++ MTFileTranslate/CHotUpdateDataStruct.h
@@ -25,7 +25,7 @@ typedef struct _FileTransferInfo{//文件传输信息
 typedef struct _FileUpdateInfo{
     int nTotal;             //文件总个数
     int nIndex;             //当前文件序号
-    float dProcess;         //传输进度
+    double dProcess;        //传输进度
     char cHandleFlag[32];   //服务端使用的标识参数，客户端不使用
     char cFileName[128];    //文件名
     char cFilePath[256];    //文件所在文件夹绝对路径
